Merge 22024 - Use GetMessagePos instead of GetCursorPos to avoid unwanted detached tabs.
The problem with GetCursorPos is that it queries the mouse position right now,
which means that the values may not be consistent with the values reported via
windows messages.  GetMessagePos on the other hand returns the mouse position
as last reported via the message loop.

With this change, I'm unable to reproduce unwanted detached tabs even under
heavy system load.

R=ben
BUG=2993
TEST=While your system is heavily loaded, click on a tab, release the mouse
button and then move the mouse downward.  Do this as quickly as you can, and
the tab should stay put.

Review URL: http://codereview.chromium.org/160364

TBR=darin@chromium.org

Review URL: http://codereview.chromium.org/160368