Linux: Fix Valgrind error in VisitedLinkEventsTest.Coalescense.
I introduced the error in r22193; it shows up when there's no X
server to connect to.  I get a "GLib-GObject-CRITICAL **:
g_object_get: assertion `G_IS_OBJECT (object)' failed" message with
this change, but I'm not sure if it's from my code or if it was
already there, and the tests still pass.

TEST=ran VisitedLinkEventsTest.Coalescense under Valgrind with DISPLAY=:6

Review URL: http://codereview.chromium.org/159753