GTK Themes: Theme labels in LocationBarViewGtk.
Labels in LocationBarViewGtk should be black even when a dark GTK theme is used
when we're not in GTK theme mode. Likewise, when we're using a GTK theme, we
should have the background and border color of the "Search X for:" box use
colors from the theme instead of having hard coded ones.

Review URL: http://codereview.chromium.org/159798