Implement a couple layoutTestController methods and rebaseline http/tests/misc/favicon-loads-with-images-disabled.html.
Implement layoutTestController.disableImageLoading() and put
in a placeholder for layoutTestController.setIconDatabaseEnabled()
(because we don't use the icon database).

These methods are used by
http/tests/misc/favicon-loads-with-images-disabled.html, but the
test still fails because our resource loader output is different.
Go ahead and rebaseline for the difference.

BUG=18355

Review URL: http://codereview.chromium.org/159820