Fixing potential memory leak in BookmarkCodec::DecodeNode.
Coverity finds a condition where this leaks. We are at the moment not using it in such a way that it leaks, but that might of course change. It looks to me like the API is not intended to be used in such a way anyway, so we can just check for it and return false.

BUG=15350
TEST=Decoding bookmarks should work as before.

Review URL: http://codereview.chromium.org/159075