Buff out Windows & Tabs API, fix a few doc bugs.
This adds callback types, event types, and descriptions for the tabs & windows api.

All defined types are loaded first in preparing the template data, so that types can cross reference across apis.

Also, additional commentary is added to content_scripts.html doc about communication with the embedding page of a content_script.

A check is now made in build.py for the case of the doc page rendering timing out (which is typically a clue that page rendering went wrong and should be inspected with chrome).

Review URL: http://codereview.chromium.org/165063