Merge 21982 - Retry requests that has timed out in BufferedDataSource
Retries request if the read request is not fulfilled in 5 seconds and at most 3 times. This will handle the case when
a user pause a movie and starts it in a later time when the
connection to the server has already timed out.

TEST=BufferedDataSourceTest.*
BUG=17479

done

Review URL: http://codereview.chromium.org/159382

TBR=hclam@google.com

Review URL: http://codereview.chromium.org/164199