linux: fix race in TransportDIB passing
There was a race condition when the renderer passes a TransportDIB to the plugin
process. It is possible that the renderer destroys the TransportDIB before the
plugin handles the message, in which case the plugin will fail to map the shm
handle.

This also cleans up a couple of things.

BUG=18320

Review URL: http://codereview.chromium.org/164226