Fix potential crash in Browser::OpenURLAtIndex.
This is a bit of a Coverity nit, but if disposition
SUPPRESS_OPEN is ever passed in to this function,
it will crash because we try to call Focus on a
new_contents pointer that is null. I verified this
by hard coding the disposition as we open URLs.

BUG=17101
TEST=No significant change.

Review URL: http://codereview.chromium.org/155773