A WebApplicationCacheHost api change allow 'foreign' entries to restart the current frame navigation.
If a document has a <html manifest='foo'> attribute that doesn't correspond with the manifest url of the cache
from which the doc was loaded, thats a foreign entry. The entry should be marked as such, and the frame
navigation restarted. The next time around, the foreign entry won't be picked up (because its foreign).

TEST=none
BUG=none


Review URL: http://codereview.chromium.org/164570