Linux: more interactive tests porting.
The most noteworthy change here is the implementation of SendMouseMove() and SendMouseClick() in ui_controls. I've combed the interwebs and I don't think it's possible to figure out the GdkWindow that is showing for a given (x,y) coordinate pair (except perhaps by delving into X), so we have to just send clicks to wherever the pointer lies. This is unfortunate in that it means we have to move the pointer, wait for it to get where it's going, and only then make the click. But on the bright side there's this super helpful function called gdk_display_warp_pointer() which makes moving the mouse a breeze.

BUG=19076

Review URL: http://codereview.chromium.org/174113