Merge 24493 - Tab strip should not resize on close (Vista).
When you close a tab using the mouse, we synthesize a fake WM_MOUSEMOVE message so that the close button that slides under the mouse will highlight. We now use GetMessagePos to read the position of the cursor and SendMessage to dispatch it, otherwise the receiving GetMessagePos call will not calculate the coordinates correctly (coordinates will be relative to the Chrome main window top left point, not the monitor's top left point, causing the tabstrip to resize).

BUG=18832
TEST=On Vista, open a bunch of tabs, then start closing the tab in the middle using the mouse. Don't move the mouse at all while doing this and observe that the tab strip doesn't resize. Also, the close button on the tab that slides under the mouse should be highlighted.

Review URL: http://codereview.chromium.org/173433

TBR=finnur@chromium.org

Review URL: http://codereview.chromium.org/174549