Added notice for blocked items to the popup container
Left a TODO to rename the class to something more general but
it is preferable to do a rename separate from code changes, it
will make things easier to sync and be quicker since that change
is relatively trivial.

The exact look may be changed later but I'll attach screenshots
to this issue to show the initial implementation.

Also left as future enhancement is the possibility of limiting
the size of the notification list in cases where a lot of hosts
are affected. The most extreme example I found was some Gizmodo
page with 36 domains using cookies.

BUG=16932
TEST=none

Review URL: http://codereview.chromium.org/179013