Squash the bookmark bar into the toolbar by 4 pixels.
On windows they do this by having the two views overlap. We can't so instead move the bottom 4 pixels of padding from the toolbar widget into the bookmark bar widget. When the bookmark bar is closed/hidden, it still shows those 4 pixels. When it opens, it is able to make use of those 4 pixels. Since we get 4 extra pixels, we can make the overall height of the bookmark bar 4 pixels less.

BUG=18570

Review URL: http://codereview.chromium.org/178029