POSIX: Have pref service save paths as UTF8.
We were passing the raw path string to StringValue, but
StringValue expects strings to be UTF8.  There's only
a DCHECK to enforce this, so I imagine it was working
silently before.

This will break for users that don't currently have UTF8
file names, but it will now work going forward.

Review URL: http://codereview.chromium.org/201027