Make hunspell.gyp refer to icu.gyp instead of icu38.gyp.
This is in preparation for http://codereview.chromium.org/172031/
(the final step in upgrading ICU to 4.2.1)

If another hunspell change gets landed AND pulled in before CL 172031 is landed, this has to be reverted. Better would be to hold off hunspell changes until CL 172031 is landed.

BUG=8198 (http://crbug.com/8198)
TEST='gclient runhooks --force' works fine after pulling in this change via DEPS update.
TBR=mhm

Review URL: http://codereview.chromium.org/194045