Rebaseline js layout tests that depend on the output strings.
This is a compromised solution. This should be fixed in V8
so V8 outputs the same results as JSC. A bug was filed in
V8: http://code.google.com/p/v8/issues/detail?id=438

Before this is fix, we create local baselines for these
tests from upstream and remove the chromium's versions, which
are basically the same tests.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/201094