Expectations for 3 perf systems on morejs and startup.
Set the perf expectations on linux, mac, and windows
perf systems for the morejs and warm startup tests.
The variance assumptions might require fine-tuning on
the master to get right.  We might be flaky to begin
with, or may want to turn off the variance reporting
altogether.

BUG=18597
TEST=build.chromium.org perf slaves stay green
Review URL: http://codereview.chromium.org/193083