Second attempt at the new syncer thread impl, now with less crashes! Previous one at http://codereview.chromium.org/214033.
I had local edits that resulted in initializing the CommandLine for syncapi,
but didn't have them as part of the patch, so this was causing a crash whenever
SyncerThreadFactory::Create was called.  The only diff here is the call to
CommandLine::Init in syncapi.cc.  This effectively means you can't change the
syncer thread impl on linux (we init an empty command line there), but this is
OK.  Once we link statically we won't need to do this.

TEST=ProfileSyncServiceTest, SyncerThreadTest, SyncerThreadWithSyncerTest

Review URL: http://codereview.chromium.org/250001