Mac: Fix crash while dragging a tab alone in a window.
Specifically, the crash could be repro'ed as follows (it also occurred under other circumstances as well): Have a window with one or more tabs in it. Create a new window with a single tab. Drag the tab in the new window, and while dragging right-click (and release).

This patch also prevents right-clicks while dragging from stopping the drag.

BUG=23591
TEST=See above.

Review URL: http://codereview.chromium.org/242139