Revert r24580, the revert of r20675 (that is, reintroduce r20675) and port http://codereview.chromium.org/200123 to the 195 branch.
Make the invisible items in the system menu of a few Indian language versions of Chrome. 

|owner_drawn| is currently set if the UI font family or the UI font size is overriden. However, this leads menu items to be invisible in the system menu. This was not the case in Chrome 2.0, but somehow the way 'owner_drawn' menu is drawn has changed and we have this regression. 

As a work around for the branch (3.0 train left, but we can merge this to 3.0.1), this CL sets |owner_drawn| to false for the system menu even if the UI font or the UI font size is overriden. 

BUG=18042
(http://crbug.com/18042 ) 
TEST=Run Chrome with '--lang=hi' command line flag on Windows. Make sure that the translated menu items in the system menu are visible (e.g. 'New Tab', 'Task Manager') although a bit too small to read while the font size in the page/wrench menu is reasonably large. 


Review URL: http://codereview.chromium.org/245033