Fix mail regression in GateKeeper.
M-A reported that GateKeeper was still sending email on
manually triggered builds or if the tree was already
closed.  Here's a rewritten isInterestingStep() that
exits earlier in hopes this fixes the problem, and the
logic around when we exit earlier should be clearer.

BUG=none
TEST=verify I can start my local master with the
gatekeeper enabled
Review URL: http://codereview.chromium.org/242159