Linux: remove the concept of physical memory from the task manager.
Physical memory isn't a concept on Linux. We have private pages and
proportional set size.

This patch matches the task manager up with about:memory. 'Private
memory' gets you the number of private pages and 'Shared memory' is
the PSS (except on systems with old kernels that don't have PSS, where
it'll be zero).

http://codereview.chromium.org/242096