Add about:memory link to task manager.
Since about:memory DCHECK()s atm, the link is not completely hooked up.

In the xib, I've added a square button with height 14 and the new HyperlinkCell, contained in a GTMWidthBasedResizer or how it's called. I made sure the baseline of link on the left and button on the right is at the same height.

BUG=17989,13156
TEST=Open MainMenu.xib, click "Enable" for view->dev->taskman. Build&run chrome. Task manager should contain link.

Review URL: http://codereview.chromium.org/255018