More gypbuild infrustructure.
No idea if this is the best structure or the correct
options but I have to start somewhere. If you have
better ideas speakup or start editing :-D

If there is a way to get gpy to list the targets
please tell me so I can call it.

Review URL: http://codereview.chromium.org/242068