Ensure that the correct URL and title is displayed in the host browser if the browser tab is attached to an existing external tab. This can happen if a script running in chrome invokes window.open which opens a popup window. To ensure that the title and URL display correctly we send over the navigation state changed IPC from Chrome when we receive an IPC from ChromeFrame indicating that it connected to the external tab.
I also made a change to the function which sends over the 
navigation state to chrome to ensure that we set the title 
to the URL if it is empty. This mimics Chrome and the other 
browser's behavior.

This should fix http://code.google.com/p/chromium/issues/detail?id=24024

Bug=24024

Review URL: http://codereview.chromium.org/274035