Seeking to keyframes only for theora
FFmpeg seeking code for theora has no way to determine if the frame is keyframe
or not, so there's no way for us to seek to the exactly the keyframe. But we
can flag that we are looking for a keyframe after a flush, and thus discard all
frames until we reach a keyframe. The outcome will be the same. This patch flags
we are performing a flush after avcodec_flush_buffers(), and
avcodec_decode_video will only return with a frame when a keyframe is reached.

Original CL (LGTMed): http://codereview.chromium.org/268034

Review URL: http://codereview.chromium.org/271106