Tracks the amount of memory SQLite is using in the browser process
This patch adds support for SQLite memory use tracking in the task manager and enables it for the browser process.  SQLite can get into the 10s of MB of memory pretty easily and it is nice to see exactly how much of a 'fat' browser process it is responsible for.

SQLite is also used in the renderer process for web databases, but since those are behind a flag and off by default I haven't added support for them yet.

TEST=none
BUG=none

Review URL: http://codereview.chromium.org/276009