Remove an overzealous version check from keystone_install.sh.
The existing version check compared the previously-installed version's
KSVersion with the newly-installed version's KSVersion after completing the
update.  In a case where a previous update attempt was interrupted after
the new KSVersion was set, this check could fail on a subsequent run of the
update script.

BUG=none
TEST=keystone_install_test.sh
Review URL: http://codereview.chromium.org/287006