Classify more types of input as queries.
This adds an oracle that can determine if a hostname is RFC 1738 compliant; if we get a hostname that's not an IP address and not compliant here, we almost certainly have a query.

This also rejects "ports" that don't consist of a number between 0 and 65535 (e.g. port "abc" or port "123456").

BUG=18817
TEST=The string "foo: bar" (with quotes) should default to search, not navigate, in the address bar.
Review URL: http://codereview.chromium.org/271116