First fix to minimize copying of image data.
This is the first of multiple patches that clean up handling of memory
regarding images. Previously, the code did several memcpy()s or
equivalents. This change:

- Creates an abstract interface RefCountedMemory which provides access to the
  front() of a memory range and the size() of it. It is a RefCountedThreadSafe.
- Adds a RefCountedStaticMemory class which isa RefCountedMemory.
- Pushes RefCountedBytes up into base/ from chrome/ and make it conform to
  RefCountedMemory.
- Have ResourceBundle return RefCountedStaticMemory to the mmaped() or DLL
  loaded resources instead of memcpy()ing them.
- General cleanups to minimize copies in constructing RefCountedBytes.
- Use the above consistent interface in the BrowserThemeProvider, along with
  special casing the loading of the new tab page background.

This patch is mostly cleanups and there should only be a slight performance
gain if any. Most of the real speedups should come in subsequent patches.

BUG=http://crbug.com/24493
TEST=Slightly faster on Perf bot; does not introduce crashes.

Review URL: http://codereview.chromium.org/288005