Cleanup the FlipDelegate API a bit in prep for fixing upload.
Document the API for FlipDelegate in flip_session.h.
Remove the method OnCancel, since it was not needed.
Add the method OnBodySent, which is not yet used but will
be.

FlipNetworkTransaction is a FlipDelegate, so the changes
there are just to reflect the new API.

BUG=none
TEST=none
Review URL: http://codereview.chromium.org/339047