linux: work around "flash does not register any clicks" bug
I had a flash of inspiration!  Setting GDK_NATIVE_WINDOWS in the
environment doesn't work because we don't propagate the setting down
into the plugin subprocess.  I confirmed that setting it manually
fixes it.  That is, on this computer I previously had the no-clicks
bug and with this change the problem is solved.

Note: there are multiple confusing levels to this bug; if
nspluginwrapper gets involved I still encounter the problem.
But this definitely fixes the problem for me.

BUG=20654

Review URL: http://codereview.chromium.org/384059