Revert part of r30340: urllib -> urllib2
A version of urllib2 is present in buildbot production
that doesn't support urllib2.urlencode.  This keeps
GateKeeper from being able to remotely close the tree.
Revert this part of r30340 for now since we're not
yet using the timeout capability in urllib2.  Also,
timeouts-as-cause of the automatic tree closure issues
is doubted as the real problem.  See review comments
at http://codereview.chromium.org/342002.

BUG=none
TEST=GateKeeper can close tree again
Review URL: http://codereview.chromium.org/393003