Hide seccomp sandbox calls from non-x86 machines
The seccomp sandbox is not built on ARM as it contains x86 specific code.
However, there are a pair of calls to seccomp sandbox functions in the zygote
initialisation, this hides those calls behind the ARCH_CPU_X86_FAMILY define.

Without this the ARM build does not link.

Review URL: http://codereview.chromium.org/402009