Merge 32240 - Ensure that the refcount on InternalGetCommandRequest stays nonzero through a PostTask
The problem was that BaseSessionService::ScheduleGetLastSessionCommands() was posting a task with a InternalGetCommandsRequest* request parameter by calling NewRunnableMethod(.., &SessionBackend::ReadLastSessionCommands, request).  SessnionBackend::ReadLastSessionCommands takes one parameter of type scoped_refptr<InternalGetCommandsRequest>.  However, NewRunnableMethod was matching the template because an InternalGetCommandsRequest* is implicitly convertable to a scoped_refptr<InternalGetCommandsRequest> but it was not actually creating the scoped_refptr<> (and thus bumping the refcount) until the task was dispatched.  By this time the refcount on the InternalGetCommandsRequest had already dropped to zero, leading to memory corruption.

This fixes the problem by passing a scoped_refptr<...> in to NewRunnableMethod() to ensure that it is copied and that the refcount stays up.

TEST=covered by TabRestoreUITest.RestoreIntoSameWindow  caused very intermittend failures locally
BUG=27191

Review URL: http://codereview.chromium.org/404012

TBR=jamesr@chromium.org
Review URL: http://codereview.chromium.org/408005