Fix unitialized memory access in workers.
The primary issue was that OnDestroy didn't change the entangled port
to have its entangled port be none.

A secondary issues that came up is that in very rare circumstances (like a crash
happening early in a worker process), it seemed like it may be possible that one
of the message ports may think it is entangled and the other half may not, so
the Erase method guards against this.

Also, some code was added to verify the internal structure before running code
and after.

BUG=27839
TEST=valgrind on linux running ui tests, specifically WorkerTest.WorkerFastLayoutTests.

Review URL: http://codereview.chromium.org/402106