Disable the replacement of U+005C (backslash) with Yen sign in documents encoded in Japanese encodings.
This is for the 249 branch. Note that this fork gets much smaller because the upstream kinda accepted
my proposal to fix http://crbug.com/3799 (Compare this with http://codereview.chromium.org/174528). 


BUG=9696
TEST=1. Go to any EUC-JP/Shift_JIS encoded web pages with backslash (e.g. LaTeX source, C/C++/Perl/Python source pages with character ecapes, CSS stylesheets 
with character escape, etc) and copy'n'paste a code snippet to oher programs. Backslashs should remain intact instead of being turned to Yen signs.
One webkit layout test fails as a result because the upstream has a test for this, but it shouldn't matter)
TBR=laforge
Review URL: http://codereview.chromium.org/440032