Mac: restore window size/position when hiding bookmark bar or download shelf.
This applies in particular when the bookmark bar or download shelf causes the
window to grow at the top (due to there not being enough space at the bottom of
the screen). Note that resizing or moving reset the state (* for moving it's a
bit more subtle), so hiding afterwards will always shrink the window from the
bottom.

BUG=29192
TEST=(1) Navigate to a normal (non-NTP) page and hide bookmark bar; move window close to the bottom of the screen (keeping the window fully on-screen); Shift-Cmd-B to show the bookmark bar (which show grow maybe a bit downwards and then upwards); Shift-Cmd-B to hide it again; it should shrink from the top and from the bottom, restoring its original size/position. (2) Do the same but move the window away from the bottom of the screen after the first Shift-Cmd-B; now upon hiding it should only shrink at the bottom. (3) Check that if the window is far enough away from the bottom of the screen to begin with, then growing/shrinking only occurs at the bottom. (4) Do (1)-(3) for the download shelf (download something to show the shelf; then hide it by closing it).

Review URL: http://codereview.chromium.org/465024