Mac: implement "Search the Web..." (Opt-Cmd-F).
This focuses the Omnibox with '?' in the contents with the text insertion point immediately after (and no selection), unless the Omnibox already has something beginning with '?' in which case it selects the text after the '?'. This is what
Ctrl-K does on other platforms. On the Mac, we get a menu item (under Edit -> Find), and the Mac-standard shortcut of Opt-Cmd-F (users can set Cmd-K as a short cut in System Preferences if they really want).

Changes to MainMenu.xib: added ^IDS_EDIT_SEARCH_WEB_MAC menu item ("Search the Web...", tag 39002, Opt-Cmd-F as keyboard shortcut, sending -commandDispatch: to first responder) as first item in Find submenu, followed by a separator.

BUG=29501
TEST=Check that this feature has the behaviour described above (using both the menu item and using Opt-Cmd-F), with (1) nothing in the Omnibox to begin with, (2) a random URL in the Omnibox, and (3) "?<some search term(s)>". Check that the Omnibox is properly focused (so set focus to various places, and make sure that the Omnibox gets focused).

Review URL: http://codereview.chromium.org/460111