Force failing CHECK/DCHECKs to end the process on Windows.
The current implementation of Windows-only hook simply calls FAIL.
According to http://code.google.com/p/googletest/issues/list?cursor=234
it is a misuse of FAIL macros (it may not shut down the process).

As a result, it was impossible to write death tests on windows.

BUG=24885
Review URL: http://codereview.chromium.org/482001