[Mac] Attempt to fix ref-after-free problems with NSTableView dataSource.
Clearing dataSource in the source's -dealloc seems generally good.
HungRendererController having a local copy of the titles is in case
the table view is redrawn as part of closing the window (this can
happen if the window is closed while the user is mousing on the
table, and perhaps at other times).

BUG=29558
TEST=Crash noted in bug stops happening.

Review URL: http://codereview.chromium.org/469016