Factor out the resource generation actions for Linux and FreeBSD.
Create a new target 'packed_resources' that can be used for building
resource files for Linux and FreeBSD, without building the chrome
binary.

The new target is useful for test programs that depend on the
resource files. For example, view_examples on Linux depends on
chrome.pak to be present in the product directory. Hence we
needed to build "chrome" before "view_examples" just for
resource files, which is not great.

Note that I tried to unify the actions with Mac, but failed
as Mac needs to build the mac bundle as part of chrome_dll
library.

TEST=try --clobber; locally clean-built chrome on linux.
BUG=29969

Review URL: http://codereview.chromium.org/486006