Fix browser crash in extension system.
When an external extension is removed from the json file
(for example by a 3rd party installer) after the user
has explicitly uninstalled it, Chrome would start crashing
and crashing.

This is now fixed. Also added a couple of unit tests for
this case (for each external provider) and cleaned up one
of those unit tests to make it match what the other was
doing.

BUG=http://crbug.com/30505
TEST=Covered by unit test.

Review URL: http://codereview.chromium.org/504016