Upload renderer histograms to browser after a page loads
WARNING: We need to watch the perf charts when this lands to be
sure that it doesn't cause a regression.

Since there is no finalizer for a renderer, and some renderers
are just terminated with no warning, the only way to get
consistent page data is to send it to the browser after
each page loads.  That is the bulk of this change.

I also fixed up a few nits relating to making this work
in single process mode.

BUG=16495
r=raman.tenneti
Review URL: http://codereview.chromium.org/515072