Fix a bug of unnecessary "Confirm form resubmission" dialog.
If a URL is requested by the POST method and then the page is
redirected to another URL, NavigationEntry::has_post_data_ is not
cleared correctly with the prior code. So "Confirm form resubmission"
dialog is shown when a user asks reloading the redirected page.


TEST=Add a unit test
BUG=21245

Review URL: http://codereview.chromium.org/524057