Unload and BeforeUnload handlers would not get invoked within pages inside ChromeFrame. This was because the ExternalTabContainer would close the underlying tab contents in its OnDestroy handler without invoking unload handlers on the page if any. In Chrome the TabStripModel invokes the RunUnloadListenerBeforeClosing on the delegate which is implemented by the browser. We need to mimic this in the external tab container.
To achieve this we now have a static helper function RunUnloadListenerBeforeClosingHelper in the browser
class which is called by the ExternalTabContainer and the Browser::RunUnloadListenerBeforeClosing function.

The ExternalTabContainer also needs to wait for the unload handlers on the page to return before it returns
control back to the host browser. To achieve this we enter a nested modal loop which is exited when the
underlying tab is closed.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=31853

Bug=31853
Test=Covered by ChromeFrame unit test.

Review URL: http://codereview.chromium.org/543183