Renabling the FullTabModeIE_ChromeFrameUnloadEventTest ChromeFrame test. This test failed at times on the builder due to a deadlock between the urlmon worker thread in IE and the ui thread. This would occur when the active document instance was going away while the worker thread was waiting for a COM call to be marshaled to the UI thread which expects the thread to pump messages.
The only interface which needs to be marshaled to the worker thread is the IBindCtx interface. Turns out
that there is nothing of note in this object, except the bind options. The rest of the properties queried
for all fail initially and then the caller register its own variants of these properties.

Fix is to create a new IBindCtx interface and copy over the BIND_OPTs to this interface.

Fixes bug http://code.google.com/p/chromium/issues/detail?id=34246


Bug=34246

Review URL: http://codereview.chromium.org/566026