Revert 38677 - Merge 38043  Merge 37799  Modify the toast to hit systemlevel installs
This is somewhat tricky. setup.exe is executed by SYSTEM and
as such we cannot run the experiment. So what we do is that
if there is an interactive user logged in at the moment of
chrome upgrade we relaunch setup in her context, from then
on things proceed as in the previous case.

Other enhancements
 New experiment codes TKxx and TLxx (99% and 1% as usual)
 New experiment group TK40 and TL40 : An active user

BUG=32474
TEST=see bug
TBR=laforge@chromium.org

Review URL: http://codereview.chromium.org/596068