Requesting C++ readability review for these classes and unit tests.
The arena allocator is designed for optimized allocation and
deallocation of small temporary objects, and is used by the
augmentable red-black tree. In a subsequent checkin, the red-black
tree will be augmented into an interval tree, which is a key data
structure for efficient implementation of one of the algorithms in the
forthcoming code in this directory.

Review URL: http://codereview.chromium.org/596093