Add user/account roles, and more.
Add user and account roles to aebuild.  Add an Account
model that is associated with users.  Update the app to
use a middleware layer that installs the current user
account in a static variable within the Account model.

Add a Config model to track the settings of the app,
including the allowed hosts that can submit and the
submit password (currently unused).

Add a Project model to keep track of all the different
projects that have submitted to us.

Add a manage interface to allow updating the Config
model.  Refactor the templates to allow reusing the
sign in/out header and links across all pages.

Update the views module to pass more variables into
templates by default.

Update the submit method to use the Config settings.

BUG=34084
TEST=login/logout works on dev, manage link appears for
manage users, settings can be updated by manage users