[Mac] Implement HTML5 database items in "Cookies and Other Data" dialog.
Code changes to support a new variant of CocoaCookieTreeNode that displays the information about web databases.

Changes to Cookies.xib: Added a new NSView "Database" in the existing information summary view for items selected in the cookie tree. This Database view provides the details for selected items that are web databases, and is in the same implement + layout style as the existing "Cookie" and "Local Storage" views. Added a key "databaseDescription" to the CocoaCookieTreeNode to support the display of the new node type.

TEST=manual testing in the "Cookies and Other Data" dialog
BUG=http://crbug.com/35191
Patch from Dan Clifford <danno@google.com>.

Review URL: http://codereview.chromium.org/596058