When ChromeFrame switches to IE on receiving the OnHttpEquiv notification indicating the presence of a meta tag indicating that the page is to be rendered in Chrome, we check if the notification is received for a site rendered in an IFrame to ensure that we don't switch in this case. This code is not reliable and we end up assuming that a top level navigation is actually an embedded frame.
Attempting another approach to detect the dummy IWebBrowser2 iframe instances which mshtml creates for top level
navigations which results in the IE to Chrome switch not working reliably.

The location for these dummy browsers is set to about:blank. We now check for the same and ignore these.
I have added comments in the code indicating why we are doing this and the cases where it could break.
It should be fine for most common cases as this code is only hit when the page has a meta tag. Will revisit
this.


This fixes http://code.google.com/p/chromium/issues/detail?id=36825

Bug=36825
Test=Covered by unit test.

Review URL: http://codereview.chromium.org/911003