Make SDCH experiment significant on the dev channel population
The experiment was reduced to a low level to avoid impacing much of
the standard release population.  Now that we've adjusted the definition
of the page-load-time metrics, I'd like to see the results of the
experiment, and I need a larger participation rate to be visible on the
dev channel.

r=willchan
Review URL: http://codereview.chromium.org/1162001