Try to fix a compile error on ARM.
This is a speculative change, to hopefully get the ARM builder compiling again.
Lately it has been giving the error that |kHandshakeResponseMessage| is being used before it is initialized:
http://build.chromium.org/buildbot/waterfall/builders/Chromium%20Arm/builds/2639/steps/compile/logs/stdio

I don't know why it would say this, but perhaps it is getting confused by the multiple statics with the same name in this file.
My proposed change removes the static keyword, to see if it has an effect.

TBR=ukai

Review URL: http://codereview.chromium.org/1217003