Merge 41433 - NSS and NPN: send NextProtocol on resume and disable NPN without SPDY.
NPN on resume was broken symmetrically between OpenSSL and NSS (which
is probably why we didn't notice for so long). OpenSSL wasn't
expecting the message and NSS wasn't sending it.

This change sends the NextProtocol message when resuming and disables
NPN when not using SPDY.

The reason for the latter is that we have GFEs deployed which don't
expect NextProtocol when resuming and we have to wait some time before
they get replaced. Because of this, we can't have NPN enabled by
default. Instead, we'll use hostresolverrules to redirect
everything to knowngood GFEs in the mean time.

http://codereview.chromium.org/841003

TBR=agl@chromium.org
Review URL: http://codereview.chromium.org/1217008