GTK: Turn on the GTK+ integration by default.
This will regress default startup time for users by 50ms. This doesn't matter
because this is what the users want--I've lost count of the number bloggers,
people on twitter, and people in our IRC channel complaining about the lack of
GTK+ integration...when it's just not on by default.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/1442001