Move enable/disable storage to prefs.
The ProfileSyncService will manage the enabled/disabled types.  Remove
CustomizeSyncWindowView's DataTypeManager dependency; it should really
just be operating on the preferences via the ProfileSyncService.

Clarify the difference between types being enabled (meaning the user
wants to sync them) versus being advertisable (meaning we support the
datatype, and advertise that the user has the option of syncing it).

BUG=34209,38340
TEST=After syncing, open the options dialog and recustomize your
options.  Verify that the enabled types from the wizard match the
initial state of the options window.  After changing the options and
hitting "accept", observe that a breakpoint in DTM::Configure gets
triggered, with exactly the new configuration set.  Verify that the
preferences are preserved on restart, as well as across Stop
Syncing / Start Syncing.

Review URL: http://codereview.chromium.org/1556009