Send JSON-encoded version of payload.
simplejson.dumps on GAE uses a lot of CPU.  Move the
JSON dumping from GAE's status listener to the Buildbot
to work around it.

We may run into more uses of simplejson in aebuild that
we can't just offload to the master, which might lead
to more quota issues to deal with.

BUG=none
TEST=packets are sent with payload json-encoded
Review URL: http://codereview.chromium.org/1564029