Remove the last remnants of windowless_needs_set_window_ from Linux plugins
Also removes the variable now that it's unused on all platforms, the WindowlessSetWindow argument that went with it, and a completely dead clipping variable.

Fixes Windows to call WindowlessSetWindow for clip rect changes (a case that used to be covered by the now-removed WindowlessSetWindow call in Draw).

BUG=none
TEST=Linux and Windows windowless plugins should still work (especially test scrolling them partially and completely off screen).

Review URL: http://codereview.chromium.org/1225009