Merge 44419 - Too many audio hangs Leopard
BUG=30242
TEST=Go to this page:
http://weston.ruter.net/projects/testcases/chromemp3audiocrash/
Click play and it won't hang.

Opening (not) too many audio streams in Leopard will cause a
deadlock in AudioQueue, freezing the browser process. Unfortunately
there isn't a way to detect that in the AudioQueue API. So this
change limits the number of streams allowed.

The limit for Leopard is set to 15 and other OSes are 50.

Review URL: http://codereview.chromium.org/1559024

TBR=hclam@chromium.org
Review URL: http://codereview.chromium.org/1733002