Add a first run bubble to show after installing an extension app.
This is similar to what we show when installing a page action or
browser action, albeit with a slightly different text.

Also disable the install warning for extension apps, for the time
being.

And finally, in tab_strip.cc I accidentally checked in a comment,
which I had commented out during testing. It was supposed to be
on (related to my last cl).

BUG=None
TEST=Install an extension app, notice a first run bubble instead
of a installation warning. The first run bubble should have a
checkbox allowing you to create a desktop shortcut.

Review URL: http://codereview.chromium.org/1693008