Queue and task fixes, plus a memcache fix.
Limit event processing to 50 events per task run.
Ensures the task has time to queue the console task.

Order processing of events by id rather than created
to ensure events are processed in order.

Reduce the queue rates for event and console queues
to 1/s.

Add a revqueue and set its rate to 10/s.

Use memcache.set instead of memcache.add when loading
the config.  Ensure we update the memcached copy of
config when we change it in manage.

BUG=none
TEST=tasks run at a stable rate, rev tasks are queued
successfully, memcache errors don't appear in the logs,
'build not found' errors are not found in the logs
Review URL: http://codereview.chromium.org/1697007