SPDY: Fix Alternate-Protocol.
(1) In DoInitConnection() we do the existing spdy session check.  If it exists there, then we assuem it exists in DoSpdySendRequest().  Unfortunately, we didn't do the same check.  Use a member variable to store the HostPortPair.
(2) In DoInitConnection(), we used the scheme://urlhost:urlport as the connection group.  With Alternate-Protocol, we used the scheme://urlhost:urlport even though we were connecting to a different port, with a different protocol (TLS).  This means we would mix conflicting sockets in the ClientSocketPool.  I fix this by dropping scheme://, since it's unnecessary, and would cause us not to share SSL sockets in different connection groups (since the specified scheme might be http://, but due to Alternate-Protocol, we actually do an SSL connect).  I also don't use the urlhost:urlport, but use the host:port that we actually connect to.

TODO(willchan):
Fix Alternate-Protocol so it works properly with proxies.  I need to change CONNECT for http proxies and patch the SOCKs connects.

Review URL: http://codereview.chromium.org/1755005