Ground work for making appcache work in workers.
Put in place a noop impl of createApplicationCacheHost() on all WebCommonWorkerClient subclasses. That factory method is soon to be added to the WebKit::WebCommonWorkerClient  class as a pure virtual. Adding these methods in advance so it builds when that pure virtual shows up later.

BUG=39368
TEST=none

Review URL: http://codereview.chromium.org/1729015