Mac BookmarkBubbleController updateBookmarkNode crash.
It is possible that due to model change the bookmark bubble window has been closed
but the popup is still showing and able to notify the controller of a folder change.
We ignore the sender in this case.

A better fix would be to close the popup programatically, but this is not a cocoa-friendly
thing to do.  I recommend this fix for M5, and for M6 we revisit to close the popup.

BUG=41257, 42766
TEST=Manual testing according to but report.  Requires two machines.  Will only crash under 10.6.

Review URL: http://codereview.chromium.org/1739015