GTK: Fix toolbar internal-padding.
The widget is called "GtkToolbar", not "GtkToolBar". Capitalization
matters. This could have been a simple one character fix, but after looking at
it, I've set the internal-padding to 1 pixel and wrapped the "Other Bookmarks"
button so that it has one pixel of padding above and below it to
match. Otherwise, it can draw into BrowserAction area.

BUG=20495
TEST=Bookmark bar doesn't look ridiculously constrained on Radiance or Ambiance GTK+ theme.

Review URL: http://codereview.chromium.org/1843002